ISSUE #3738 - rename due date components and add disabled state
